WIP getting meal tables to render after addition or deletion of food.
