+++ foods.html
@@ -12,6 +12,7 @@
   <div class='new-and-filter-container'>
     <div class='label'>Filter by Name</div>
     <input type='text' id='search-term'>
+    <button id='redirect-to-diary'>Go To Diary</button></br>
   </div>
 
   <table id='food-table' class="filterable">
