+++ lib/foods.js
@@ -1,79 +1,6 @@
 const helperFile = require('./helper.js')
+const mealFile = require('./meals.js')
 
-const getFoods = () =>  {
-  $('#food-table tr > td').remove()
-  fetch(`${helperFile.url}api/v1/foods`)
-  .then((response) => helperFile.handleResponse(response))
-  .then((foods) => getEachFood(foods))
-  .catch((error) => console.error ({ error }))
-}
-
-const getEachFood = (foodData) => {
-  foodData.forEach((food) => {
-    renderFood(food)
-  })
-}
-
-const getPage = ()  =>  {
-  let urlParts = document.URL.split('/')
-  let lastSegment = urlParts.pop() || urlParts.pop()
-  return lastSegment
-}
-
-const renderFood = (food, meal=null) => {
-  let page = getPage()
-  if(page == "foods.html") {
-    renderFoodTableContent(food)
-  } else {
-      if (meal != null) {
-        renderMealTableContent(food, meal)
-      }
-      else {
-        renderAvailableFoodTableContent(food)
-      }
-  }
-  $('.delete-icon').on("click", deleteFood)
-  $('.delete-from-meal-icon').on("click", deleteMealFood)
-}
-
-const renderFoodTableContent = (food) =>{
-  $('#food-table tr:first').after(`<tr data-food=${food.id}>
-    <td contenteditable="true" class='name'>${food.name.toUpperCase()}</td>
-    <td contenteditable="true" class='calories'>${food.calories}</td>
-    <td class='delete-icon'><img src='assets/delete.png' alt='delete'>
-  </tr>`)
-}
-
-const renderMealTableContent = (food, meal) => {
-  let tableName = meal
-  $(`#${tableName} tr:first`).after(`<tr data-food=${food.id}>
-    <td contenteditable="true" class='name'>${food.name.toUpperCase()}</td>
-    <td contenteditable="true" class='calories'>${food.calories}</td>
-    <td class='delete-from-meal-icon'><img src='assets/delete.png' alt='delete'>
-  </tr>`)
-}
-
-const renderAvailableFoodTableContent = (food) => {
-  $('#available-food-table tr:first').after(`<tr data-food=${food.id}>
-    <td contenteditable="true" class='name'>${food.name.toUpperCase()}</td>
-    <td contenteditable="true" class='calories'>${food.calories}</td>
-    <td class='checkbox'><input type="checkbox"></td>
-  </tr>`)
-}
-
-const addNewFood = (event) => {
-  event.preventDefault()
-  let newName = $("#new-food-name").val().toUpperCase(),
-      newCalories = $("#new-food-calorie-count").val(),
-      body = {food: {name: newName, calories: newCalories}}
-  if (!(getFoodNames()).includes(newName)){
-    makeRequest('POST', '', body)
-    .then((response) => helperFile.handleResponse(response))
-    .catch((error) => console.error({error}))
-    $("#new-food-form").trigger('reset')
-    setTimeout(getFoods, 300)
-  }
-}
 
 const getFoodNames = () => {
 return $('#food-table').find('tr').map((i, row) => {
@@ -81,18 +8,6 @@ return $('#food-table').find('tr').map((i, row) => {
       return $(row).find('.name').text().toUpperCase()
     }
   }).toArray()
-}
-
-
-$("#new-food-button").on("click", addNewFood)
-
-const deleteFood = (event) => {
-  event.preventDefault()
-  let row = event.target.closest('tr')
-  let id = row.dataset.food
-  makeRequest('DELETE', id)
-  .catch((error) => console.error({error}))
-  setTimeout(getFoods, 300)
 }
 
 $("#food-table").keypress((event) => {
@@ -141,22 +56,19 @@ const filterFoods = () => {
   })
 }
 
-$('#search-term').on('keyup', filterFoods)
-
-const deleteMealFood = (event) => {
-  event.preventDefault()
-  let foodId = event.target.closest('tr').dataset.food,
-      mealId = event.target.closest('table').dataset.meal
-  fetch(`${helperFile.url}api/v1/meals/${mealId}/foods/${foodId}`, {
-    method: 'DELETE',
-    headers: {
-      "Content-Type": "application/json"
-    }
-  })
-  .then((response) => console.log(response))
-  .catch((error) => console.error({error}))
-  
+const renderFoodTableContent = (food) =>{
+  $('#food-table tr:first').after(`<tr data-food=${food.id}>
+    <td contenteditable="true" class='name'>${food.name.toUpperCase()}</td>
+    <td contenteditable="true" class='calories'>${food.calories}</td>
+    <td class='delete-icon'><img src='assets/delete.png' alt='delete'>
+  </tr>`)
 }
 
+$('#search-term').on('keyup', filterFoods)
+
+$("#redirect-to-diary").on( "click", function() {
+  window.location.href = "http://localhost:8080/webpack-dev-server/"
+})
+
 export * from './foods.js'
-module.exports = {getFoods, renderFood, getPage}
+module.exports = { getFoodNames, renderFoodTableContent, makeRequest }
