+++ lib/helper.js
@@ -14,4 +14,4 @@ const handleResponse = (response) => {
   })
 }
 
-module.exports = {handleResponse, url}
+module.exports = { handleResponse, url }
