+++ lib/index.js
@@ -1,8 +1,11 @@
 import * as Style from '../main.css'
 import * as Foods from './foods.js'
+import * as Meals from './meals.js'
 
 const foodFile = require('./foods.js')
+const mealFile = require('./meals.js')
 const helperFile = require('./helper.js')
+
 var dir;
 var originalTable;
 
@@ -35,76 +38,6 @@ const addIdstoTables = (meal) => {
   mealTable.dataset.meal = mealId
 }
 
-
-const renderMeals= (meal) => {
-  let tableName = meal.name
-  meal.foods.forEach((food) => {
-    foodFile.renderFood(food, tableName)
-  })
-  calculateTotalCalories(tableName)
-  calculateCalorieSummary()
-}
-
-const calculateTotalCalories = (tableName)  =>  {
-  let mealTable = document.getElementById(tableName)
-  let calories = mealTable.getElementsByClassName("calories")
-  let sum = 0
-
-  for (let i = 0; i < calories.length; i++) {
-    let value = calories[i].innerHTML
-    if(!isNaN(value) && value.length != 0) {
-        sum += parseFloat(value)
-    }
-  }
-  let totalCaloriesDisplay = mealTable.getElementsByClassName('total-calories')[0]
-  totalCaloriesDisplay.innerHTML = sum
-  calculateRemainingCalories(mealTable, totalCaloriesDisplay.innerHTML)
-}
-
-const calculateRemainingCalories = (table, totalCalories) => {
-  let startingCalories;
-
-  switch (table.id) {
-    case 'Snack':
-      startingCalories = 200
-      break;
-    case 'Breakfast':
-      startingCalories = 400
-      break;
-    case 'Lunch':
-      startingCalories = 600
-      break;
-    case 'Dinner':
-      startingCalories = 800
-  }
-  let remainingCaloriesDisplay = table.getElementsByClassName('remaining-calories')[0]
-  remainingCaloriesDisplay.innerHTML = startingCalories - totalCalories
-  styleRemainingCalories(remainingCaloriesDisplay)
-}
-
-const styleRemainingCalories = (remainingCalories) =>  {
-  if (remainingCalories.innerHTML >= 0) {
-    remainingCalories.parentElement.style.backgroundColor = "#79B473"
-  } else {
-    remainingCalories.parentElement.style.backgroundColor = "#CD5C5C"
-  }
-}
-
-const calculateCalorieSummary = () => {
-  let totalSum = 0
-  $('.total-calories').each(function(calorieCount)  {
-    let value = $(this).context.innerHTML
-    let convertToNumber = parseInt(value)
-    totalSum += convertToNumber
-  })
-
-  let goalCalories = $('#goal-calories-total').text()
-  let remainingCalories = $('#remaining-calories-total')
-  $('#total-calories-total').text(totalSum)
-  remainingCalories.text(goalCalories - totalSum)
-  styleRemainingCalories(remainingCalories[0])
-}
-
 $("#redirect-to-foods").on( "click", function() {
   window.location.href = "http://localhost:8080/webpack-dev-server/foods.html"
 })
@@ -116,12 +49,6 @@ $("#meal-buttons").on("click", function(event) {
   getMealFoods(mealId)
   $('#available-food-table').find('input[type=checkbox]:checked').removeAttr('checked')
 })
-
-const getMealFoods = (mealId) => {
-  fetch(`${helperFile.url}api/v1/meals/${mealId}/foods`)
-  .then(response => helperFile.handleResponse(response))
-  .then(meal => renderMeals(meal))
-}
 
 const addFoodToMeal = (mealId) => {
   let FoodIds = getCheckedFoods()
@@ -210,9 +137,103 @@ $("th.calories").hover(function() {
   })
 
 $(document).ready(() => {
-  foodFile.getFoods()
-  let page = foodFile.getPage()
+  getFoods()
+  let page = getPage()
   if (page != 'foods.html'){
     getMeals()
   }
 })
+
+// Dependent upon Food and Meal files
+
+const getPage = ()  =>  {
+  let urlParts = document.URL.split('/')
+  let lastSegment = urlParts.pop() || urlParts.pop()
+  return lastSegment
+}
+
+const renderFood = (food, meal=null) => {
+  let page = getPage()
+  if(page == "foods.html") {
+    foodFile.renderFoodTableContent(food)
+  } else {
+      if (meal != null) {
+        mealFile.renderMealTableContent(food, meal)
+      }
+      else {
+        mealFile.renderAvailableFoodTableContent(food)
+      }
+  }
+  $('.delete-icon').on("click", deleteFood)
+  $('.delete-from-meal-icon').on("click", deleteMealFood)
+}
+
+const renderMeals= (meal) => {
+  let tableName = meal.name
+  meal.foods.forEach((food) => {
+    renderFood(food, tableName)
+  })
+  mealFile.calculateTotalCalories(tableName)
+  mealFile.calculateCalorieSummary()
+}
+
+const getFoods = () =>  {
+  $('#food-table tr > td').remove()
+  fetch(`${helperFile.url}api/v1/foods`)
+  .then((response) => helperFile.handleResponse(response))
+  .then((foods) => getEachFood(foods))
+  .catch((error) => console.error ({ error }))
+}
+
+const getEachFood = (foodData) => {
+  foodData.forEach((food) => {
+    renderFood(food)
+  })
+}
+
+const addNewFood = (event) => {
+  event.preventDefault()
+  let newName = $("#new-food-name").val().toUpperCase(),
+      newCalories = $("#new-food-calorie-count").val(),
+      body = {food: {name: newName, calories: newCalories}}
+  if (!(foodFile.getFoodNames()).includes(newName)){
+    foodFile.makeRequest('POST', '', body)
+    .then((response) => helperFile.handleResponse(response))
+    .catch((error) => console.error({error}))
+    $("#new-food-form").trigger('reset')
+    setTimeout(getFoods, 300)
+  }
+}
+
+const deleteFood = (event) => {
+  event.preventDefault()
+  let row = event.target.closest('tr')
+  let id = row.dataset.food
+  foodFile.makeRequest('DELETE', id)
+  .catch((error) => console.error({error}))
+  setTimeout(getFoods, 300)
+}
+
+const getMealFoods = (mealId) => {
+  let numId = parseInt(mealId)
+  $('[data-meal=' + numId + '] > .meal-food-info').remove()
+  fetch(`${helperFile.url}api/v1/meals/${numId}/foods`)
+  .then(response => helperFile.handleResponse(response))
+  .then(meal => renderMeals(meal))
+}
+
+const deleteMealFood = (event) => {
+  event.preventDefault()
+  let foodId = event.target.closest('tr').dataset.food,
+      mealId = event.target.closest('table').dataset.meal
+  fetch(`${helperFile.url}api/v1/meals/${mealId}/foods/${foodId}`, {
+    method: 'DELETE',
+    headers: {
+      "Content-Type": "application/json"
+    }
+  })
+  .catch((error) => console.error({error}))
+  setTimeout(getMealFoods(mealId), 300)
+}
+
+$("#new-food-button").on("click", addNewFood)
